Move TaggerTrainer to the new tensorflow::trainer module
